+++ mayan/apps/mayan_statistics/apps.py
@@ -1,16 +1,15 @@
 from django.utils.translation import ugettext_lazy as _
 
 from mayan.apps.common.apps import MayanAppConfig
-from mayan.apps.common.menus import menu_object, menu_secondary, menu_tools
+from mayan.apps.common.menus import (
+    menu_list_facet, menu_object, menu_return, menu_tools
+)
 from mayan.apps.navigation.classes import SourceColumn
 
 from .classes import StatisticLineChart, StatisticNamespace
 from .links import (
-    link_statistic_namespace_detail,
-    link_statistic_namespace_list,
-    link_statistic_queue,
-    link_statistic_detail,
-    link_statistics
+    link_statistic_namespace_detail, link_statistic_namespace_list,
+    link_statistic_queue, link_statistic_detail, link_statistics
 )
 
 
@@ -45,18 +44,18 @@ class StatisticsApp(MayanAppConfig):
             label=_('Last update'), source=StatisticLineChart
         )
 
+        menu_list_facet.bind_links(
+            links=(link_statistic_namespace_detail,),
+            sources=(StatisticNamespace,)
+        )
         menu_object.bind_links(
             links=(link_statistic_detail, link_statistic_queue),
             sources=(StatisticLineChart,)
         )
-        menu_object.bind_links(
-            links=(link_statistic_namespace_detail,),
+        menu_return.bind_links(
+            links=(link_statistic_namespace_list,),
             sources=(StatisticNamespace,)
         )
-        menu_secondary.bind_links(
-            links=(link_statistic_namespace_list,),
-            sources=(
-                StatisticNamespace, 'statistics:statistic_namespace_list'
-            )
+        menu_tools.bind_links(
+            links=(link_statistics,)
         )
-        menu_tools.bind_links(links=(link_statistics,))