+++ HISTORY.rst
@@ -127,6 +127,7 @@
 - Show cache partitions and partition file totals. This helps determine how
   effective is a cache maximum size value by showing how many objects and
   files the cache size limit is able to yield.
+- Improve statistics navigation.
 
 4.3.1 (2022-08-21)
 ==================